Move `run_from_json` into `flaps_api post`
