+++ internal/command/machine/flaps_api/post.go
@@ -1,4 +1,4 @@
-package machine
+package flaps_api
 
 import (
 	"context"
@@ -24,15 +24,21 @@ import (
 	"github.com/superfly/flyctl/internal/watch"
 )
 
-func newRunFromJson() *cobra.Command {
+var sharedFlags = flag.Set{
+	flag.App(),
+	flag.AppConfig(),
+	flag.Detach(),
+}
+
+func newPost() *cobra.Command {
 	const (
 		short = "Run a machine from a JSON configuration file"
 		long  = short + "\n"
 
-		usage = "run_from_json <json_file> [command]"
+		usage = "post <json_file> [command]"
 	)
 
-	cmd := command.New(usage, short, long, runMachineRunFromJson,
+	cmd := command.New(usage, short, long, runPost,
 		command.RequireSession,
 		command.LoadAppNameIfPresent,
 	)
@@ -40,11 +46,6 @@ func newRunFromJson() *cobra.Command {
 	flag.Add(
 		cmd,
 		flag.Region(),
-		// deprecated in favor of `flyctl machine update`
-		flag.String{
-			Name:        "id",
-			Description: "Machine ID, if previously known",
-		},
 		flag.String{
 			Name:        "name",
 			Shorthand:   "n",
@@ -92,7 +93,7 @@ func loadMachineJson(ctx context.Context, path string, input *api.LaunchMachineI
 	return nil
 }
 
-func runMachineRunFromJson(ctx context.Context) error {
+func runPost(ctx context.Context) error {
 	var (
 		appName  = app.NameFromContext(ctx)
 		client   = client.FromContext(ctx).API()
@@ -112,14 +113,14 @@ func runMachineRunFromJson(ctx context.Context) error {
 	}
 
 	if appName == "" {
-		app, err = createApp(ctx, "Running a machine without specifying an app will create one for you, is this what you want?", "", client)
+		app, err = mach.CreateApp(ctx, "Running a machine without specifying an app will create one for you, is this what you want?", "", client)
 		if err != nil {
 			return err
 		}
 	} else {
 		app, err = client.GetAppCompact(ctx, appName)
 		if err != nil && strings.Contains(err.Error(), "Could not find App") {
-			app, err = createApp(ctx, fmt.Sprintf("App '%s' does not exist, would you like to create it?", appName), appName, client)
+			app, err = mach.CreateApp(ctx, fmt.Sprintf("App '%s' does not exist, would you like to create it?", appName), appName, client)
 			if app == nil {
 				return nil
 			}
@@ -142,12 +143,8 @@ func runMachineRunFromJson(ctx context.Context) error {
 		return fmt.Errorf("the app %s uses an earlier version of the platform that does not support machines", app.Name)
 	}
 
-	input.Config, err = determineMachineConfig(ctx, *input.Config, app, flag.FirstArg(ctx), input.Region, false)
-	if err != nil {
-		return err
-	}
-
 	input.AppID = app.Name
+	// Support a few command-line overrides, but this shouldn't be a clone of `machine run`
 	if name := flag.GetString(ctx, "name"); name != "" {
 		input.Name = name
 	}
