+++ internal/machine/create_app.go
@@ -1,0 +1,64 @@
+package machine
+
+import (
+	"context"
+
+	"github.com/superfly/flyctl/api"
+	"github.com/superfly/flyctl/internal/prompt"
+)
+
+func CreateApp(ctx context.Context, message, name string, client *api.Client) (*api.AppCompact, error) {
+	confirm, err := prompt.Confirm(ctx, message)
+	if err != nil {
+		return nil, err
+	}
+
+	if !confirm {
+		return nil, nil
+	}
+
+	org, err := prompt.Org(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	if name == "" {
+		name, err = selectAppName(ctx)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	input := api.CreateAppInput{
+		Name:           name,
+		OrganizationID: org.ID,
+	}
+
+	app, err := client.CreateApp(ctx, input)
+	if err != nil {
+		return nil, err
+	}
+
+	return &api.AppCompact{
+		ID:       app.ID,
+		Name:     app.Name,
+		Status:   app.Status,
+		Deployed: app.Deployed,
+		Hostname: app.Hostname,
+		AppURL:   app.AppURL,
+		Organization: &api.OrganizationBasic{
+			ID:   app.Organization.ID,
+			Slug: app.Organization.Slug,
+		},
+	}, nil
+}
+
+func selectAppName(ctx context.Context) (name string, err error) {
+	const msg = "App Name:"
+
+	if err = prompt.String(ctx, &name, msg, "", false); prompt.IsNonInteractive(err) {
+		err = prompt.NonInteractiveError("name argument or flag must be specified when not running interactively")
+	}
+
+	return
+}
